requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-groq
faiss-cpu
python-dotenv
pydantic
